document-processor/requirements.txt
flask
flask-cors
python-dotenv
pytesseract
Pillow
